(1 Sep 2019) Updated the if-else options for the computation and added delivery total in hours
I updated the calculations if-else statement because I was missing an entire range of numbers that are possible to get during calculation from the range. The numbers between 60 and 100 now have a computational replacement value that is to be calculated. If a number falls within this range, the number will be subtracted from 100 to give the equivalent for the minutes that remain within the range of the time period. 

For example, if 70 remained in the range, the new method will subtract 70 from 100. This leaves 30 remaining. 30/60 = 0.5, which means 0.5 hours or half an hour.

In addition, there is a function that is added to display the total hours worked immediately after placing an end time to an entry.